feat: all tests delete runing

+++ src/infra/db/mongodb/product/product-mongo-repository.ts
@@ -24,27 +24,15 @@ export class ProductMongoRepository
   async create(payload: AddProductModel): Promise<ProductModel> {
     try {
       const productCollection = await MongoHelper.getCollection('products');
-      const categorieCollection = await MongoHelper.getCollection('categories');
-      const category = await categorieCollection.findOne({
-        _id: new ObjectId(payload.categoryId),
-      });
-      if (!category) {
-        throw new BadRequestException(
-          `Category with ${payload.categoryId} id not found.`,
-        );
-      }
-      const result = await (
-        await productCollection
-      ).insertOne({ ...payload, category });
+
+      const result = await productCollection.insertOne(payload);
+
       const product = await productCollection.findOne({
         _id: result.insertedId,
       });
 
       return MongoHelper.map(product);
     } catch (error) {
-      if (error instanceof BadRequestException) {
-        throw new BadRequestException(error.message);
-      }
       throw new InternalServerErrorException(error.message);
     }
   }
@@ -118,17 +106,18 @@ export class ProductMongoRepository
         }),
       );
 
+      if (!product) {
+        throw new BadRequestException(`Product with id ${id} not found.`);
+      }
+
       await productCollection.deleteOne({
         _id: new ObjectId(id),
       });
 
       return product;
     } catch (error) {
-      if (
-        (error.message =
-          "Cannot destructure property '_id' of 'collection' as it is null.")
-      ) {
-        throw new BadRequestException(`Product with id ${id} not found.`);
+      if (error instanceof BadRequestException) {
+        throw new BadRequestException(error.message);
       }
       throw new InternalServerErrorException(error.message);
     }
